% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dataElement}
\alias{dataElement}
\title{dataElement}
\description{
R6 class for holding data and related info (metadata). It holds a single
element of data which can be anything, though the original intent was for it
to hold data.frame's and additional meta data. The object is capable of also
holding related functions that can be initiated from the object itself via
the 'do' function. These functions are contained in their own
'dataFunctions' R6 class object.

A data element can have an id (integer) and/or a name (character vector).
These are optional.
}
\note{
apart from ... the parameter data = self is also passed on in order
to give the function access to the data in the dataElement object. Be
careful changing the data parameter as it might affect the calling
dataElement.

if the new info contains a column named "id",
then this will be used to set the id of the object.
Also: the actual (private) info_ will not contain
the "id" column
}
\examples{
theData <- dataElement$new(
data = datasets::penguins,
id = 1,
name = "penguins",
info = list(source = "datasets")
)
theData$info
theData$data |> head()
# use of dataFunctions
thePlot <- function(data) {
plot(data$data)
}
theSummary <- function(data) {
  summary(data$data)
}
theData$dos <- dataFunctions$new(functions = list(theSummary, thePlot),
                                 names = c("summary","plot"))
theData$dos
theData$do("summary")[1:6, 3:4]
theData$do("plot")

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{id}}{gets & sets the id of the object}

\item{\code{name}}{gets & sets the name of the object}

\item{\code{info}}{gets & sets the info of the object}

\item{\code{data}}{get & sets the data of the object}

\item{\code{dos}}{provides access to dataFunctions object associated with the
dataElement object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dataElement-new}{\code{dataElement$new()}}
\item \href{#method-dataElement-print}{\code{dataElement$print()}}
\item \href{#method-dataElement-do}{\code{dataElement$do()}}
\item \href{#method-dataElement-clone}{\code{dataElement$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataElement-new"></a>}}
\if{latex}{\out{\hypertarget{method-dataElement-new}{}}}
\subsection{Method \code{new()}}{
initializes the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataElement$new(
  id = NA,
  name = "",
  info = NA,
  infoClass = as.data.frame,
  data = NA,
  do = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{integer value that sets the id of the dataElement}

\item{\code{name}}{character value that sets the name of the dataElement}

\item{\code{info}}{specifies the meta data of the dataElement. This is best
either a list or a data.frame.}

\item{\code{infoClass}}{function, specifies what to use to transform the info
argument into (usuallt a data.frame)}

\item{\code{data}}{data to be stored in the dataElement (usually a data.frame)}

\item{\code{do}}{a dataFunctions object that sets which functions can be used
with the dataElement}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataElement-print"></a>}}
\if{latex}{\out{\hypertarget{method-dataElement-print}{}}}
\subsection{Method \code{print()}}{
For printing purposes: prints the info object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataElement$print()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{no arguments, the function takes care of printing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataElement-do"></a>}}
\if{latex}{\out{\hypertarget{method-dataElement-do}{}}}
\subsection{Method \code{do()}}{
Function that starts one of the functions in the associated
dataFunctions object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataElement$do(whichFunction, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{whichFunction}}{character vector specifying which function to call.
Note: function needs to be present in the dataFunction object associated
with the dataElement object}

\item{\code{...}}{for passing on parameters to the function to be called}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataElement-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dataElement-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataElement$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
