% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{dataFunctions}
\alias{dataFunctions}
\title{R6 class for holding a set of functions}
\description{
Basic R6 class that holds functions, that can be used to dynamically add or
change functions for another object. See manual for full example.
}
\note{
if the whichFunction argument is not present in the names of the
functions in the object the 'empty function
}
\examples{
dataF <- dataFunctions$new()
dataF
dataF$functions
dataF$names
dataF$add(theFunction = mean, theName = "mean")
dataF$add(theFunction = median, theName = "median")
dataF
dataF$do("mean")(datasets::mtcars$mpg)
dataF$do("median")(datasets::mtcars$mpg)
dataF <- dataFunctions$new(
  functions = list(mean, plot),
  names = c("mean", "plot")
)
dataF
dataF$do("plot")(datasets::mtcars)

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{length}}{returns the number of functions present in the object, read only}

\item{\code{functions}}{gets & sets the (list of) functions present in the object.
If data is put into the object it is checked (whether it is a function).}

\item{\code{names}}{gets & sets the names of all the functions}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dataFunctions-new}{\code{dataFunctions$new()}}
\item \href{#method-dataFunctions-print}{\code{dataFunctions$print()}}
\item \href{#method-dataFunctions-add}{\code{dataFunctions$add()}}
\item \href{#method-dataFunctions-do}{\code{dataFunctions$do()}}
\item \href{#method-dataFunctions-clone}{\code{dataFunctions$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataFunctions-new"></a>}}
\if{latex}{\out{\hypertarget{method-dataFunctions-new}{}}}
\subsection{Method \code{new()}}{
initializes the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataFunctions$new(functions = list(), names = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{functions}}{list of functions to be added upon initialization. Please note
that the class of list elements is checked!}

\item{\code{names}}{names to be used for the functions in the 'functions' argument}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataFunctions-print"></a>}}
\if{latex}{\out{\hypertarget{method-dataFunctions-print}{}}}
\subsection{Method \code{print()}}{
For printing purposes: prints the names of the contained functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataFunctions$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{no arguments, the function takes care of printing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataFunctions-add"></a>}}
\if{latex}{\out{\hypertarget{method-dataFunctions-add}{}}}
\subsection{Method \code{add()}}{
adds a function to the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataFunctions$add(theFunction = NA, theName = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theFunction}}{the function to be stored in the object}

\item{\code{theName}}{name with which to call function (does not have to be the name
of the actual function).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataFunctions-do"></a>}}
\if{latex}{\out{\hypertarget{method-dataFunctions-do}{}}}
\subsection{Method \code{do()}}{
retrieves a function 'stored' in the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataFunctions$do(whichFunction = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{whichFunction}}{name of the function to be retrieved}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataFunctions-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dataFunctions-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataFunctions$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
