% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dataInfo}
\alias{dataInfo}
\title{dataInfo}
\value{
logical vector

NA or an integer vector
}
\description{
R6 class for holding a set of dataElement objects. The original intent was
for all dataElements to be the same. If they are not, it should still work,
though extra care needs to be take to the 'dataFunctions' called via do().
}
\note{
no arguments, the function takes care of printing

all functions must take the parameters: data which refers to the
dataInfo itself, index & id to be able to work on the calling dataInfo
object. Please note that to prevent dataInfo data modification, the
dataInfo object should be cloned (deep = TRUE).

for debugging the code & solving issues. Should not be manually set
w/o (good) reason. May be removed in the future
}
\examples{
tElement <- list(
 dataElement$new(
   name = "Test 001",
   data = datasets::mtcars,
   info = data.frame(name = c("mtcars"))
 ),
 dataElement$new(
   name = "Test 002",
   data = datasets::iris,
   info = data.frame(name = c("iris"))
 ),
 dataElement$new(
   name = "Test 003",
   data = datasets::volcano,
   info = data.frame(name = c("volcano"))
 )
)
tElement
tInfo <- dataInfo$new(data = tElement)
tInfo
tInfo$length
tInfo$add(
 data = list(dataElement$new(
   name = "Penguins",
   data = datasets::penguins,
   info = list(name = "penguins", datasource = "datasets")
 ))
)
tInfo

# different ways of referring to the actual data
tInfo$item(id = 1) |> head()
tInfo$data[[1]] |> head()
tInfo$data$`Test 001` |> head()
tInfo$raw[[1]]
tInfo$raw[[1]]$data |> head()
tInfo$item.list(id = 1:2)
# register & use functions
dimensions <- function(data, index = 1, id = NA) {
 result <- list()
 for (counter in index) {
  result[[length(result) + 1]] <- dim(data$item(index = counter))
  names(result)[length(result)] <- data$names[counter]
 }
 return(result)
}
dimensions(data = tInfo)
dimensions(data = tInfo, index = 1:tInfo$length)

plotData <- function(data, index = 1, id = NA) {
  plot(data$item(index = index, id = id))
}
plotData(tInfo)
tInfo$dos <- dataFunctions$new(
functions = list(dimensions, plotData),
 names = c("dimensions","plot")
)

tInfo$do(whichFunction = "dimensions", index = 1)
tInfo$do(whichFunction = "dimensions", index = 1:tInfo$length)
tInfo$do("plot")
tInfo$do("plot", id = 2)
tInfo$do("plot", id = 3)
tInfo$do("plot", id = 4)

}
\section{Super class}{
\code{\link[dataInfo:dataElement]{dataInfo::dataElement}} -> \code{dataInfo}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{info}}{gets & sets the info part of the dataElements in the dataInfo
object}

\item{\code{data}}{gets & sets the data of the objects in the dataElements list}

\item{\code{raw}}{provides direct access to the objects in the dataElements list}

\item{\code{length}}{returns the number of dataElements in the object (read only).}

\item{\code{index}}{set & gets the current value of the (private) index (info_)
field}

\item{\code{ids}}{gets & sets the id's of the dataElements in the dataInfo object}

\item{\code{names}}{gets & sets the names of the dataElements in the dataInfo
object}

\item{\code{empty}}{logical vector, determines if the object is (still) empty
(read only)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dataInfo-new}{\code{dataInfo$new()}}
\item \href{#method-dataInfo-print}{\code{dataInfo$print()}}
\item \href{#method-dataInfo-save}{\code{dataInfo$save()}}
\item \href{#method-dataInfo-load}{\code{dataInfo$load()}}
\item \href{#method-dataInfo-validIndex}{\code{dataInfo$validIndex()}}
\item \href{#method-dataInfo-indexFromId}{\code{dataInfo$indexFromId()}}
\item \href{#method-dataInfo-idFromIndex}{\code{dataInfo$idFromIndex()}}
\item \href{#method-dataInfo-getIndex}{\code{dataInfo$getIndex()}}
\item \href{#method-dataInfo-add}{\code{dataInfo$add()}}
\item \href{#method-dataInfo-delete}{\code{dataInfo$delete()}}
\item \href{#method-dataInfo-item}{\code{dataInfo$item()}}
\item \href{#method-dataInfo-item.list}{\code{dataInfo$item.list()}}
\item \href{#method-dataInfo-do}{\code{dataInfo$do()}}
\item \href{#method-dataInfo-clone}{\code{dataInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataInfo-new"></a>}}
\if{latex}{\out{\hypertarget{method-dataInfo-new}{}}}
\subsection{Method \code{new()}}{
create a new dataInfo object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataInfo$new(name = "", data = list(), clone = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character vector, name of the dataInfo object}

\item{\code{data}}{list of dataElement objects to be added to new object}

\item{\code{clone}}{logical value, defines whether the dataElements to be added
are to be 'cloned'. Default is TRUE to prevent odd situations, option
may be removed in the future}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new 'dataInfo' object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataInfo-print"></a>}}
\if{latex}{\out{\hypertarget{method-dataInfo-print}{}}}
\subsection{Method \code{print()}}{
For printing purposes: prints the info of the different dataElements as a
tibble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataInfo$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{no arguments, the function takes care of printing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataInfo-save"></a>}}
\if{latex}{\out{\hypertarget{method-dataInfo-save}{}}}
\subsection{Method \code{save()}}{
saves the private data_ list (of dataElement). By default
the file format is '.rds' (via \code{link[base]{saveRDS}})
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataInfo$save(path = "", filename = "", overwrite = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{path where to place file, should end in '/'}

\item{\code{filename}}{prefix to the file name to be added. If not specified,
then the object's name will be used}

\item{\code{overwrite}}{logical vector that defines what to do if there is
already a file with the file name to be used}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
logical vector TRUE if save was successful (otherwise FALSE)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataInfo-load"></a>}}
\if{latex}{\out{\hypertarget{method-dataInfo-load}{}}}
\subsection{Method \code{load()}}{
loads the specified file into the private data_ field. The
data in the file should be a list of dataElement objects. By default
data is loaded from '.rds' files via the \code{link[base]{readRDS}}
function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataInfo$load(path = "", filename = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{path where to place file, should end in '/'}

\item{\code{filename}}{prefix to the file name to be added}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
logical vector TRUE if load successful,
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataInfo-validIndex"></a>}}
\if{latex}{\out{\hypertarget{method-dataInfo-validIndex}{}}}
\subsection{Method \code{validIndex()}}{
checks if the index number provided is larger than 0 and
lower than or equal to the length of the list of dataElement objects in
the object. This can be used as a check to prevent out-of-range index
values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataInfo$validIndex(index = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{integer vector to be checked}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataInfo-indexFromId"></a>}}
\if{latex}{\out{\hypertarget{method-dataInfo-indexFromId}{}}}
\subsection{Method \code{indexFromId()}}{
finds index number of the dataElement in the list which
has a certain id
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataInfo$indexFromId(id = NA, na.rm = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{integer vector: the id for which the row number is to be found}

\item{\code{na.rm}}{logical vector determines if NA's should be removed from the
result}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NA or integer vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataInfo-idFromIndex"></a>}}
\if{latex}{\out{\hypertarget{method-dataInfo-idFromIndex}{}}}
\subsection{Method \code{idFromIndex()}}{
finds the id in the specified dataElement in the object list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataInfo$idFromIndex(index = NA, na.rm = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{integer vector: the index of the object, which id is to be
returned}

\item{\code{na.rm}}{logical vector determines if NA's should be removed from the
result}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NA or integer vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataInfo-getIndex"></a>}}
\if{latex}{\out{\hypertarget{method-dataInfo-getIndex}{}}}
\subsection{Method \code{getIndex()}}{
a lot of functions using the object will either give an
index or an id when working with specific items in the dataElements-list.
If both are provided, the id is used to get the index. If no id is
provided, then the index parameter is returned.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataInfo$getIndex(index = 1, id = NA, na.rm = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{default is 1, only returned when id is NA}

\item{\code{id}}{default is NA, if not, then used preferentially to return the
index of the object which has it}

\item{\code{na.rm}}{logical value determining if NA values are to be removed from
the result. Default is FALSE}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataInfo-add"></a>}}
\if{latex}{\out{\hypertarget{method-dataInfo-add}{}}}
\subsection{Method \code{add()}}{
add dataElement(s) to the dataInfo object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataInfo$add(dataElements = list(), clone = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataElements}}{function or list of dataElement objects to be added
to the dataInfo object. If it's a function, then it will be executing
inside the add() function and should result in a list of dataElement
objects.}

\item{\code{clone}}{logical value, defines whether the dataElements to be added
are to be 'cloned'. Default is TRUE to prevent odd situations, option
may be removed in the future}

\item{\code{...}}{to pass on parameters to the parameter data (only if that is a
function)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataInfo-delete"></a>}}
\if{latex}{\out{\hypertarget{method-dataInfo-delete}{}}}
\subsection{Method \code{delete()}}{
deletes a dataElement item from the (internal) list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataInfo$delete(index = NULL, id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{integer vector: indexes of the dataElement list to be deleted.
Ignored if the argument 'id' is specified}

\item{\code{id}}{id's that will be deleted from the dataElement list.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataInfo-item"></a>}}
\if{latex}{\out{\hypertarget{method-dataInfo-item}{}}}
\subsection{Method \code{item()}}{
retrieves single dataElement from the dataElement list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataInfo$item(index = NA, id = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{index of the item in the dataElement list to be retrieved.
Ignored if parameter 'id' is specified}

\item{\code{id}}{id of item that needs to be retrieved from the dataElement list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataInfo-item.list"></a>}}
\if{latex}{\out{\hypertarget{method-dataInfo-item.list}{}}}
\subsection{Method \code{item.list()}}{
retrieves a list of elements from the dataElement list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataInfo$item.list(index = 1:self$length, id = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{indexes of the items in the dataElement list to be retrieved.
Ignored if parameter 'id' is specified}

\item{\code{id}}{id's of items that need to be retrieved from the dataElement list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataInfo-do"></a>}}
\if{latex}{\out{\hypertarget{method-dataInfo-do}{}}}
\subsection{Method \code{do()}}{
executes one of the functions 'registered'. in the dataInfo's
'dos' object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataInfo$do(whichFunction, index = 1, id = NA, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{whichFunction}}{character vector: name of the function to be executed}

\item{\code{index}}{of the item(s) in the dataElement list which are to be be
used by the function. Ignored if parameter 'id' is specified}

\item{\code{id}}{id of dataElements which are to be used by the function}

\item{\code{...}}{for passing on additional parameters to the function to be
execute}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataInfo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dataInfo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
