% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dataList}
\alias{dataList}
\title{dataList}
\value{
dataInfo item
}
\description{
R6 Class to deal with a group of 'dataInfo' objects in an organized manner
}
\note{
no arguments, the function takes care of printing

all functions must take the parameters: data which refers to the
dataInfo object itself, index & id to be able to work on the calling
dataInfo object. Please note that to prevent dataInfo data modification,
the dataInfo object should be cloned (deep = TRUE).
}
\examples{
testList <- dataList$new(
name = "Test",
dataObjects = list(
  dataInfo$new(
  name = "cars"
  ),
  dataInfo$new(
    name = "USArrests"
  ),
  dataInfo$new(
    name = "Penguins"
  )
)
)
testList
testList$item("cars")
# add data to 'cars'
testList$item("cars")$add(
  data = list(dataElement$new(
    data = datasets::mtcars,
    info = list(source = "datasets")
  ))
)
testList
testList$item("cars")
# add data again (as example)
testList$item("cars")$add(
  data = list(dataElement$new(
    data = datasets::mtcars,
    info = list(source = "datasets", extra = "Once again")
  ))
)
testList
testList$item("cars")
testList$data$cars
testList$data$cars$item(1) |> head()
identical(testList$data$cars$item(1), testList$item("cars")$item(id = 2))
testList$data$Penguins$add(
  data = list(dataElement$new(
    data = datasets::penguins,
    info = list(source = "datasets")
  ))
)
testList
testList$info
testList$data$Penguins
testList$data$Penguins$data[[1]] |> head()
# create a function for some data analysis
findMeansetc <- function(
    data,
    index = 1,
    id = NA,
    selectedIsland = "Biscoe",
    selectedSpecies = "Adelie",
    selectedSex = c("female", "male")
) {
  data$item(index = index, id = id) |>
    dplyr::filter(
      island \%in\% selectedIsland,
      species \%in\% selectedSpecies,
      sex \%in\% selectedSex
    ) |>
    dplyr::group_by(island, species, sex) |>
    dplyr::summarize(
      meanBill_len = mean(bill_len, na.rm = TRUE),
      sdBill_len = sd(bill_len, na.rm = TRUE),
      count = dplyr::n()
    ) |>
    dplyr::ungroup()
}
# different ways of applying the function to the data
findMeansetc(data = testList$data$Penguins)
findMeansetc(
  data = testList$data$Penguins,
  selectedIsland = c("Biscoe", "Dream"),
  selectedSpecies = c("Adelie"),
  selectedSex = "male"
)
# adding the function to the dataList object
# (to the 'Penguins' dataInfo object of course!)
testList$data$Penguins$dos <- dataFunctions$new(
  functions = list(findMeansetc),
  names = "gather"
)
testList$item("Penguins")$dos
testList$item("Penguins")$do("gather", id = 1)
testList$item("Penguins")$do(
  "gather",
  id = 1,
  selectedIsland = c("Biscoe", "Dream"),
  selectedSpecies = c("Adelie"),
  selectedSex = "male"
)
testList$dos
testList$do(whichElement = "Penguins", whichFunction = "gather", id = 1)
testList$do(
  whichElement = "Penguins",
  whichFunction = "gather",
  id = 1,
  selectedIsland = c("Biscoe", "Dream"),
  selectedSpecies = c("Adelie"),
  selectedSex = "male"
)

}
\section{Super classes}{
\code{\link[dataInfo:dataElement]{dataInfo::dataElement}} -> \code{\link[dataInfo:dataInfo]{dataInfo::dataInfo}} -> \code{dataList}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ids}}{to stop calls to inherited functions. Nothing happens,
simply returns NA, read only}

\item{\code{data}}{provides direct access to the list of dataInfo items}

\item{\code{info}}{returns a dataframe showing the names, classes and number of
items in each dataInfo element in the dataList object}

\item{\code{names}}{provides direct access to names of the dataInfo items
in the dataList object}

\item{\code{dos}}{returns a list of the dataFunction names of each dataInfo
elementin the dataList object, read only}

\item{\code{classes}}{returns the (main) class of all dataInfo items in the
dataList object, read only}

\item{\code{classes.full}}{returns the full class vectors of all dataInfo items
in the dataList object, read only}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dataList-new}{\code{dataList$new()}}
\item \href{#method-dataList-print}{\code{dataList$print()}}
\item \href{#method-dataList-do}{\code{dataList$do()}}
\item \href{#method-dataList-indexFromId}{\code{dataList$indexFromId()}}
\item \href{#method-dataList-idFromIndex}{\code{dataList$idFromIndex()}}
\item \href{#method-dataList-add}{\code{dataList$add()}}
\item \href{#method-dataList-delete}{\code{dataList$delete()}}
\item \href{#method-dataList-item}{\code{dataList$item()}}
\item \href{#method-dataList-item.list}{\code{dataList$item.list()}}
\item \href{#method-dataList-clone}{\code{dataList$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="dataInfo" data-topic="dataInfo" data-id="getIndex"><a href='../../dataInfo/html/dataInfo.html#method-dataInfo-getIndex'><code>dataInfo::dataInfo$getIndex()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dataInfo" data-topic="dataInfo" data-id="load"><a href='../../dataInfo/html/dataInfo.html#method-dataInfo-load'><code>dataInfo::dataInfo$load()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dataInfo" data-topic="dataInfo" data-id="save"><a href='../../dataInfo/html/dataInfo.html#method-dataInfo-save'><code>dataInfo::dataInfo$save()</code></a></span></li>
<li><span class="pkg-link" data-pkg="dataInfo" data-topic="dataInfo" data-id="validIndex"><a href='../../dataInfo/html/dataInfo.html#method-dataInfo-validIndex'><code>dataInfo::dataInfo$validIndex()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataList-new"></a>}}
\if{latex}{\out{\hypertarget{method-dataList-new}{}}}
\subsection{Method \code{new()}}{
create a new dataList object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataList$new(name = "", dataObjects = NA, clone = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{character vector, name of the dataList object}

\item{\code{dataObjects}}{list of dataInfo objects to be added to new object}

\item{\code{clone}}{logical value, defines whether the dataElements to be added
are to be 'cloned'. Default is TRUE to prevent odd situations, option
may be removed in the future}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new 'dataList' object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataList-print"></a>}}
\if{latex}{\out{\hypertarget{method-dataList-print}{}}}
\subsection{Method \code{print()}}{
For printing purposes: prints the names of the "data" items present
and the number of data_ items in each
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataList$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{no arguments, the function takes care of printing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataList-do"></a>}}
\if{latex}{\out{\hypertarget{method-dataList-do}{}}}
\subsection{Method \code{do()}}{
executes one of the functions 'registered'. in of the dataInfo
objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataList$do(whichElement, whichFunction, index = 1, id = NA, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{whichElement}}{character vector: name of the dataInfo object in the
dataList that needs to be used}

\item{\code{whichFunction}}{character vector: name of the function to be executed.
The function should be available in the dataInfo object, otherwise NA will
be returned}

\item{\code{index}}{of the item(s) in the dataElement list (of the dataInfo object
to be used) which are to be beused by the function. Ignored if parameter
'id' is specified}

\item{\code{id}}{id of the dataElement(s) (in the dataInfo object to be used)
which is to be used by the function}

\item{\code{...}}{for passing on additional parameters to the function to be
execute}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataList-indexFromId"></a>}}
\if{latex}{\out{\hypertarget{method-dataList-indexFromId}{}}}
\subsection{Method \code{indexFromId()}}{
to stop calls to inherited functions. Nothing happens,
simply returns NA
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataList$indexFromId(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{no arguments, the function takes care of printing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataList-idFromIndex"></a>}}
\if{latex}{\out{\hypertarget{method-dataList-idFromIndex}{}}}
\subsection{Method \code{idFromIndex()}}{
to stop calls to inherited functions. Nothing happens,
simply returns NA
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataList$idFromIndex(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{no arguments, the function takes care of printing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataList-add"></a>}}
\if{latex}{\out{\hypertarget{method-dataList-add}{}}}
\subsection{Method \code{add()}}{
adds an dataInfo object to the data_ list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataList$add(dataObject, clone = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataObject}}{data object to be added, must be a descendant class
of 'dataInfo' (or 'dataInfo' class itself). dataInfo object should have
a name (not enforced at this moment)}

\item{\code{clone}}{logical value, defines whether the dataElements to be added
are to be 'cloned'. Default is TRUE to prevent odd situations, option
may be removed in the future}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataList-delete"></a>}}
\if{latex}{\out{\hypertarget{method-dataList-delete}{}}}
\subsection{Method \code{delete()}}{
deletes one of the dataInfo items in the dataList object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataList$delete(index = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{number or name of the dataInfo item to be deleted}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataList-item"></a>}}
\if{latex}{\out{\hypertarget{method-dataList-item}{}}}
\subsection{Method \code{item()}}{
retrieves one of the dataInfo items
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataList$item(index = 1, clone = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{number or name of the dataInfo item to be retrieved}

\item{\code{clone}}{logical vector specifying whether the returned item
should be a clone or not. Default = FALSE}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataList-item.list"></a>}}
\if{latex}{\out{\hypertarget{method-dataList-item.list}{}}}
\subsection{Method \code{item.list()}}{
retrieves one or more of the dataInfo items
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataList$item.list(index = 1:self$length, clone = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{numbers or names of the data item to be retrieved}

\item{\code{clone}}{logical vector specifying whether the returned items
should be clones or not. Default = FALSE}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list of dataInfo items
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataList-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dataList-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dataList$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
