% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFunctions.R
\name{emptyFunction}
\alias{emptyFunction}
\title{emptyFunction}
\usage{
emptyFunction(...)
}
\arguments{
\item{...}{can be used to pass on (named) arguments. Doesn't matter what is
passed on, nothing is used.}
}
\value{
NA
}
\description{
default function that doesn't do anything. Parameters can be
passed on,but nothing is used. Always returns NA. Essentially a 'fake' function.
}
\examples{
emptyFunction()

}
