% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctions.R
\name{readCSV}
\alias{readCSV}
\title{readCSV}
\usage{
readCSV(
  filename,
  columns = 1:2,
  columnNames = c("x", "y"),
  additionalInfo = NA,
  sep = ",",
  skip = 0,
  header = TRUE,
  ...
)
}
\arguments{
\item{filename}{the name of the file which the data are to be read from.}

\item{columns}{specifies which columns to take from the data. Can be a character or
an integer vector}

\item{columnNames}{specifies which column names to give to the data. Should be
same length as the columns argument. Applied to ALL data.frame's in the list}

\item{additionalInfo}{named list object that contains info to be put into the
info data.frame of the info objects. The intention is to put some relevant
information there, like source of the data, file names and other information
which makes it possible to recreate the data. Again, this argument is applied
to all elements in the data.frame list. If not specified (default is NA), then
only a basic info element is generated with source = "csv" and filename =
filename.}

\item{sep}{the field separator character.}

\item{skip}{integer: the number of lines of the data file to skip before
beginning to read data.}

\item{header}{a logical value indicating whether the file contains the names
of the variables as its first line. See also \link[utils]{read.csv}}

\item{...}{additonal parameters to be passed on to the \link[utils]{read.csv}}
}
\value{
a function that will generate a list of dataElement objects (length
is 1 in this case!)
}
\description{
function factory which generates a function that reads a .csv
file and returns an single element list of dataObjects. See also \link{readDataFrame}
}
\examples{
filename <- tempfile()
utils::write.csv(datasets::mtcars, file = filename, row.names = FALSE)
result <- readCSV(filename = filename)()
result[[1]]$info
result[[1]]$data |> head(10)
unlink(filename)
}
