% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctions.R
\name{readExcel}
\alias{readExcel}
\title{readExcel}
\usage{
readExcel(
  filename,
  sheet = 1,
  columns = 1:2,
  columnNames = c("x", "y"),
  rowNames = NULL,
  additionalInfo = NA,
  useReadxl = FALSE
)
}
\arguments{
\item{filename}{name of the excel file from which the data to read}

\item{sheet}{name or number of the sheet in the excel file to read data from}

\item{columns}{columns from the excel file to read}

\item{columnNames}{new names for the columns of the data read from the excel
file (length column names should be the same as the length of columns)}

\item{rowNames}{specifies which row names to give to the data.frame. Should
be same length as the number of rows of the data being read. Default is NULL}

\item{additionalInfo}{additional info to be added to the result of the read
data function. Should be named list format, default is NA}

\item{useReadxl}{logical vector. If TRUE the readxl package is used to read
the excel file, if FALSE then the XLConnect package is used. default is
FALSE}
}
\value{
a function that reads the data from the specified excel file and
returns a list of dataElement objects
}
\description{
function factory to read excel data
}
\note{
Uses the 'XLConnect' or the 'readxl' package to read the data from the
file
}
\examples{
# mtcars xlsx file from demoFiles subfolder of package XLConnect
demoExcelFile <- system.file("demoFiles/mtcars.xlsx", package = "XLConnect")
result <- readExcel(demoExcelFile)()
result[[1]]$info
result[[1]][["data"]] |> head()
result <- readExcel(demoExcelFile, rowNames = rownames(datasets::mtcars))()
result[[1]]$data |> head()
result <- readExcel(demoExcelFile, columns = 1:2, columnNames = c("x", "y"))()
result[[1]]$data |> head()

}
