% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{strReplaceAll}
\alias{strReplaceAll}
\title{strReplaceAll}
\usage{
strReplaceAll(string, pattern = NA, replacement = "")
}
\arguments{
\item{string}{one or more character vectors containing parts that need to be
replaced}

\item{pattern}{one or more character vectors that need to be replaced}

\item{replacement}{one or more character vectors that are used as replacements
This argument has to be either length one (used for all patterns) or must have
the same length as the pattern argument}
}
\value{
character vector(s)
}
\description{
extended version of str_replace_all from the stringr package.
More than one pattern and replacement can be replaced
}
\note{
patterns (and their) replacements are handled sequentially, so one after
another
}
\examples{
strReplaceAll(string = c("Hello", "Word"), pattern = "r", replacement = "rl")
strReplaceAll(string = c("Hello", "Word"), pattern = c("e","o"), replacement = "")
strReplaceAll(string = c("Hello", "Word"), pattern = c("e","o"), replacement = "")
strReplaceAll(string = c("Hello", "Word"), pattern = c("e","o"), replacement = c("a", "-"))

}
